add versioning to dotnet + couple versions
